Changed structure to atomic design
