+++ app/components/organisms/counter_screen/counter_screen.tsx
@@ -1,33 +1,15 @@
-import { BodyT1 } from '@atoms/text/text'
 import { useAppStore } from '@hooks/stores/stores'
 import { observer } from 'mobx-react-lite'
-import { Button, View } from 'react-native'
+import { View } from 'react-native'
+import React from 'react'
 import styled from 'styled-components/native'
+import { Counter } from '@molecules/counter/counter'
 
 const Wrapper = styled(View)`
   flex: 1;
   justify-content: space-between;
   align-items: center;
-`
-
-const Row = styled(View)`
-  flex: 1;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-`
-
-const IncrementButton = styled(Button)`
-  background-color: red;
-`
-
-const DecrementButton = styled(Button)`
-  background-color: orange;
-`
-
-const StyledBody = styled(BodyT1)`
-  color: black;
-  font-weight: 700;
+  background-color: white;
 `
 
 const CounterScreen = () => {
@@ -39,11 +21,7 @@ const CounterScreen = () => {
 
   return (
     <Wrapper>
-      <Row>
-        <DecrementButton title="Decrement" onPress={onPress} />
-        <StyledBody>{value}</StyledBody>
-        <IncrementButton title="Decrement" onPress={onPress} />
-      </Row>
+      <Counter value={value} increment={onPress} decrement={onPress} />
     </Wrapper>
   )
 }
