+++ app/themes/styled.d.ts
@@ -1,0 +1,6 @@
+import {} from 'styled-components/native'
+import { TStyledComponentsTheme } from './styled_components_theme'
+
+declare module 'styled-components/native' {
+  export interface DefaultTheme extends TStyledComponentsTheme {}
+}
